update telegram sender

+++ src/main/java/ru/cotarius/reminder/entity/User.java
@@ -24,7 +24,7 @@ public class User{
     private String email;
 
     @Column(name = "telegram_id")
-    private long telegramId;
+    private int telegramId;
 
     @Column(name = "username")
     private String username;
@@ -32,10 +32,8 @@ public class User{
     @Column(name = "password")
     private String password;
 
-//    @Enumerated(EnumType.STRING)
     @Column(name = "role")
     private String role;
-//    private Role role;
 
 
 }
